Rename desktop dropdown
